refactor(playground): split block visualization panel
