+++ packages/agent-playground/src/components/playground/block-visualization/block-visualization-panel/index.tsx
@@ -1,0 +1,2 @@
+export { BlockVisualizationPanel } from './block-visualization-panel';
+export type { IBlockVisualizationPanelProps } from './types';
